requirements.txt
# API and HTTP
requests>=2.28.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7

# Google Cloud
google-cloud-bigquery>=3.9.0
google-auth>=2.16.0

# Environment and configuration
python-dotenv>=1.0.0

# Date and time handling
pytz>=2023.3

# Scheduling
schedule>=1.2.0

# SSL Certificate handling
pyOpenSSL>=23.1.0
cryptography>=40.0.0

# Testing
pytest>=7.3.1
httpx>=0.24.0  # For testing FastAPI applications
